requirements.txt
streamlit==1.22.0
PyPDF2==3.0.1
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
fpdf2==2.7.4
nltk==3.8.1
python-dotenv==1.0.0